+++ viewMyTravels/VersionHistory.md
@@ -1,4 +1,5 @@
 # 0.0.2
 
-- _Marker Displays correctly on front end_
+- _Map/Marker Displays correctly._
 - _Change Location WIP_
+- _Change Location changes to new city_